Commentaire + diver fix

+++ src/Service/Command/CommandWriterService.php
@@ -15,11 +15,19 @@ class CommandWriterService
 {
     public function __construct(private EntityManagerInterface $entityManager) {
     }
+
+    /**
+     * Ajoute un $product à la Command active de $user. (Créer la Command si elle n'éxiste pas)
+     * @param Product $product
+     * @param User $user
+     * @return void
+     */
     public function addProductToCommand(Product $product, User $user): void
     {
         $commandReaderService = new CommandReaderService();
         $command = $commandReaderService->getUserActiveCommand($user);
 
+        // Créer la Command si elle n'éxiste pas
         if (!$command) {
             $command = new Command();
             $command->setUser($user);
@@ -30,6 +38,7 @@ class CommandWriterService
 
         $commandLine = $commandReaderService->getCommandLineByProduct($command, $product);
 
+        // Créer la CommandLine si elle n'éxiste pas
         if (!$commandLine) {
             $commandLine = new CommandLine();
             $commandLine->setCommand($command);
@@ -37,12 +46,19 @@ class CommandWriterService
             $commandLine->setPrice($product->getPrice());
         }
 
+        // Si la line existe ajoute 1 à la quantité
         $commandLine->setQuantity(($commandLine->getQuantity() ?? 0) + 1);
 
         $this->entityManager->persist($commandLine);
         $this->entityManager->flush();
     }
 
+    /**
+     * Valide la command si elle appartien à l'utilisateur
+     * @param Command $command
+     * @param User $user
+     * @return bool
+     */
     public function validateCommand(Command $command, User $user): bool
     {
         if($user->getId() === $command->getUser()->getId()) {
@@ -54,9 +70,17 @@ class CommandWriterService
         return false;
     }
 
-    public function deleteCommand(Command $command): void
+    /**
+     * Supprime $command si elle appartien à $user
+     * @param Command $command
+     * @param User $user
+     * @return void
+     */
+    public function deleteCommand(Command $command, User $user): void
     {
-        $this->entityManager->remove($command);
-        $this->entityManager->flush();
+        if($user->getId() === $command->getUser()->getId()) {
+            $this->entityManager->remove($command);
+            $this->entityManager->flush();
+        }
     }
 }
